requirements.txt
aiogram==3.10.0
python-dotenv>=1.0
